+++ AlchemistCode/Cards/Common/Upwell.cs
@@ -13,7 +13,7 @@ public class Upwell : AlchemistCard
 
     public Upwell() : base(1, CardType.Skill, CardRarity.Common, TargetType.Self)
     {
-        WithBlock(8, 3);
+        WithBlock(10, 3);
         WithVar("poison", 2, 0);
         WithTip(typeof(PoisonPower));
     }